Generic Pipe implementation
