+++ src/System.IO.Pipelines/src/System.IO.Pipelines.csproj
@@ -6,23 +6,31 @@
   <ItemGroup>
     <Compile Include="Properties\InternalsVisibleTo.cs" />
     <Compile Include="System\IO\Pipelines\BufferSegment.cs" />
+    <Compile Include="System\IO\Pipelines\BytePipeReaderWrapper.cs" />
+    <Compile Include="System\IO\Pipelines\BytePipeWriterWrapper.cs" />
     <Compile Include="System\IO\Pipelines\CompletionData.cs" />
     <Compile Include="System\IO\Pipelines\FlushResult.cs" />
-    <Compile Include="System\IO\Pipelines\InlineScheduler.cs" />
     <Compile Include="System\IO\Pipelines\IDuplexPipe.cs" />
-    <Compile Include="System\IO\Pipelines\Pipe.DefaultPipeReader.cs" />
-    <Compile Include="System\IO\Pipelines\Pipe.DefaultPipeWriter.cs" />
+    <Compile Include="System\IO\Pipelines\InlineScheduler.cs" />
+    <Compile Include="System\IO\Pipelines\IDuplexPipeOfT.cs" />
     <Compile Include="System\IO\Pipelines\Pipe.cs" />
+    <Compile Include="System\IO\Pipelines\PipeOfT.DefaultPipeReader.cs" />
+    <Compile Include="System\IO\Pipelines\PipeOfT.DefaultPipeWriter.cs" />
+    <Compile Include="System\IO\Pipelines\PipeOfT.cs" />
     <Compile Include="System\IO\Pipelines\PipeAwaitable.cs" />
     <Compile Include="System\IO\Pipelines\PipeCompletion.cs" />
     <Compile Include="System\IO\Pipelines\PipeCompletionCallback.cs" />
     <Compile Include="System\IO\Pipelines\PipeCompletionCallbacks.cs" />
     <Compile Include="System\IO\Pipelines\PipeOptions.cs" />
+    <Compile Include="System\IO\Pipelines\PipeOptionsOfT.cs" />
     <Compile Include="System\IO\Pipelines\PipeReader.cs" />
+    <Compile Include="System\IO\Pipelines\PipeReaderOfT.cs" />
     <Compile Include="System\IO\Pipelines\PipeOperationState.cs" />
     <Compile Include="System\IO\Pipelines\PipeScheduler.cs" />
     <Compile Include="System\IO\Pipelines\PipeWriter.cs" />
+    <Compile Include="System\IO\Pipelines\PipeWriterOfT.cs" />
     <Compile Include="System\IO\Pipelines\ReadResult.cs" />
+    <Compile Include="System\IO\Pipelines\ReadResultOfT.cs" />
     <Compile Include="System\IO\Pipelines\ResultFlags.cs" />
     <Compile Include="System\IO\Pipelines\ThrowHelper.cs" />
   </ItemGroup>
