+++ src/System.IO.Pipelines/src/System/IO/Pipelines/IDuplexPipe.cs
@@ -1,6 +1,6 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
 
 namespace System.IO.Pipelines
 {
